LUS-12628 hsm: get a valid cookie for RAoLU request
Add a way to get a valid cookie when nobody initializes
cdt_last_cookie.

RAoLU policy is allowed to queue a remove request with the
coordinator stopped. In that cases cdt_last_cookie can not be yet
initialize and the remove request can be queued with a conflicting
cookie.

This patch adds cdt_update_last_cookie() that reverses process the hsm
llog and stops at the first non-cancel action to determine the last
cookie.

Add the regression test sanity-hsm 26e.

Lustre-change: https://review.whamcloud.com/51850
Lustre-commit: 241cf3c6d08277c4a401ec8bd109274123bf9cdf
Lustre-author: Etienne AUJAMES <etienne.aujames@cea.fr>

Change-Id: Ib3c4feb560c3826cd9619511adc47c64fc17e172
WC-bug-id: LU-16235
Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>
Signed-off-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163142
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>