LUS-12632 tests: Fix usage of LNDPATH in load_lnet
LNDPATH (and LNETLND) environment variables allow one to specify a
custom path for LND kernel module, but load_lnet() would not use
LNDPATH in the call to load_module(). Thus, if a custom path was
defined then it would not be used.

Also, the default value of LNDPATH was relative to the current working
directory. It should instead use an absolute path based on LUSTRE
environment variable. This ensures load_lnet() can find LND modules
regardless of the current working directory. load_module() is updated
to detect if it is given an absolute module path.

Test-Parameters: trivial
Fixes: 118872e737 ("LUS-12525 o2ib: normalize in-kernel-ko2iblnd.ko name")
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I51de98870d1675546cc85caa510f5691f0326969
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163145
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>