LUS-12710 build: dkms in-kernel check kernel source
From dkms the default $kernel_source_dir will be
  @MODDIR@/$kernelver/build
unless otherwise specified with the --kernelsourcedir option.

Unfortunately the check for kernel headers will fail as only
the generated headers are available from the 'build' symbolic
link on SUSE.

Probe for the 'source' symbolic link if it points to a directory
with 'include/linux' proceed to prefer it when checking for
in-kernel rdma support.

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: If14cc1b71ea7bfa2cb608b7fe4f3e6450078fa0d
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163179
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>