LUS-12555 tests: Check detailed resend stats
Modify sanity-lnet health tests to check that failed_resends is
incremented correctly.

Add a test to check that successful_resends is incremented
correctly.

Modify the behavior of lnet_drop_rule_add() when the specified rate
is 2 so that the first matched message is always dropped. This allows
test_232() to work correctly. A typo is also fixed in this function
so that the correct destination NID is printed.

cleanup_health_test() would overwrite the return value from
lnetctl net del. This is fixed.

Some style fixes were made in functions touched by this patch.

Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Ibfd56fe8f0e01b1027d1295d8748b285db799e9f
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163188
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Reviewed-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>