LUS-12585 utils: Do not inherit stripe size from DOM component
When a DOM file is created with multiple components,
the size of the first component is inherited by the
other components if their sizes are not explicitly specified.

This fix ensures that the size of subsequent components is
set to the default size when no size is provided.

Lustre-change:https://review.whamcloud.com/57501
Lustre-commit: e526de5150b4f29a7ccebfd0a85a01d1e5d24f4e

Test-Parameters: testlist=sanity-dom
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: I90944aa0d10b52e28d07edb65361a0867c0e48b6
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163178
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>