LUS-12672 dkms: race on clobber and create of modules.order
DKMS builds fail occasionally with an error:

cat: /var/lib/dkms/.../build//modules.order: No such file or directory
  MODPOST /var/lib/dkms/.../build/Module.symvers

This appears to be a make bug trying where a path with //
is not understood correctly.

Remove the unnecessary injection of / in the list of SUBDIRS
to be built in when ldiskfs is not enabled.

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Icd9f98778f1d62dbfebfa4a89652579e7e72efd8
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163200
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>