LUS-12715 lustre: replace 0-length arrays with flexible arrays (2/3)
This patch replaces 0-length arrays with flexible arrays to
resolve the UBSAN array-index-out-of-bounds runtime warnings.

Most replacement of 0-length arrays with flexible arrays
requires no special handling. Simply removing the “0” in
the array declaration is sufficient.

In order to have a flexible array member in a union or alone
in a struct, it needs to be wrapped in an anonymous struct
with at least 1 named member, but that member can be empty.
This was wrapped in Linux with the DECLARE_FLEX_ARRAY() macro.

Lustre-change: 4920a934baf06a81157df90abb08225517352479
Lustre-commit: https://review.whamcloud.com/55940
Lustre-author: Jian Yu <yujian@whamcloud.com>

WC-bug-id: LU-17794
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I94c1823172aae33425f4637652bc26b72e908a91
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163226
Reviewed-by: Rajeev Mishra <rajeevm@hpe.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>