LUS-12823 ptlrpc: evictor/prolong timeout formula fix
Let's stay with the main rule - cluster is supposed to survive just
one (the worst ever) failure, not a set of failures. The current
formula covers:

- an RPC timeout;
- the 1st reconnect time out;
- the 2nd reconnect to the failover pair returns an error;
- the 3rd reconnect to the original node succeeds;

i.e. 2 failures. At the same time, The 2nd (reconnect timeout) failure
was the result of the LNET problems, which were fixed separately.

Besides that, a fix for the ping evictor timeout, a normal RPC case,
where the timeout of an ping itself was not accounted in the formula.

Besides that, a fix for the extra timeout used for obd_eviction_timer,
where net latencies for the accounted RPCs were taken not correctly.

Signed-off-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Change-Id: Ic7ccc83c5f10a8dd3382b9a337b70bc7c41a70ff
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163269
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Andriy Skulysh <andriy.skulysh@hpe.com>