LU-18983 nodemap: support obdfilter-survey with nodemap
When running obdfilter-survey case=disk, no nodemap is associated with
the export, as the echo client is setup local to the server and the
ofd is accessed directly. So allow no nodemap for a ECHO_UUID client.
When running obdfilter-survey case=network, the echo client setup on
remote node (usually client) connects to the server via the echo obd.
So the echo obd needs to add the remote nid to the nodemap.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ic8464a29e1dd02ae7e453b6a3c9484f130c01d05
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59120
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Aurelien Degremont <adegremont@nvidia.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>