LUS-12907 tests: llog zeroed record
The patch introduces recovery-small test 163 to reproduce
an EIO error while writing an llog record. This results
in partial write failures, where a section of the llog
file — possibly in the middle — is filled with zeros
instead of the expected llog data, leading to potential
misinterpretation of the llog contents.

Test-Parameters: testlist=recovery-small/163
Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I92bf0819acd34a256516ffbd93b6676ec6789cd9
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163359
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Reviewed-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Reviewed-by: Andriy Skulysh <andriy.skulysh@hpe.com>