LUS-12836 lov: Do not display lmm_stripe_offset for mdt file
The command lfs getstripe <file> displays lmm_stripe_offset
as 0 for DOM files, which can be confusing.

The lmm_stripe_offset points to the OST index where the file's
data resides. However, for DOM files, the first component is
stored on the MDT, not an OST. This discrepancy in
interpretation can lead to understanding that DOM resides
on OST 0, and in case DOM resides on MDT indext 1 it causes
user to think DOM is on MDT 0.

Added test_909() to test the change

WC-bug-id: LU-18982
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: I894aa8982c1a18f73887137c58199c1e6aa203e2
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163381
Tested-by: Jenkins Build User <nssreleng@cray.com>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>