LUS-12968 tests: lnetctl import tests
sanity-lnet test_255 contains a test of the import command, but it
was not written correctly and with the current code it silently fails.
The test has been corrected so that LNet is correctly initialized
prior to executing the import command.

sanity-lnet test_350 is added to exercise the --exec import option
using the discover and ping commands.

WC-bug-id: LU-19126
Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Iff6e7176029d5fb8a4b058aa8538586995dac9da
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163412
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>