LUS-12966 lnet: Fix CPT parsing in jt_import
jt_import() simply passes through the CPT yaml sequence to netlink,
however the kernel code expects user-space to have expanded any
specified ranges into a list of cores. i.e. the kernel understands
CPT: "[0, 1, 2, 3]"
but not
CPT: "[0-3]"

Modify jt_import() to use handle_cpt_sequence() when it encounters
a CPT sequence while processing a net sequence. handle_cpt_sequence()
is slightly modified to avoid code duplication now that it has an
additional caller.

A range like '[a-b]' is only valid if a <= b. Have
cfs_range_expr_parse() check this condition so that it can detect
invalid ranges.

While testing I found that yaml_lnet_config_ni() can segfault if
cfs_expr_list_values() is called with an empty list because it assumes
the cpt_array it passed in must be freed, but this is only the case
when the expression list is not empty. The code for emitting the CPT
sequence is moved into a new helper, lnet_yaml_emit_cpt_sequence(),
and the potential segfault is fixed by free'ing cpt_array only when
count > 0.

WC-bug-id: LU-19160
Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I874e24e6d69aea9c56d72d07494998c9a869be55
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163417
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>