LUS-12960 ptlrpc: virtual 4k bulk alignment
A 64k side cannot transfer a 64k page if the first page has
an offset, so the total size exceeds LNet_MTU. However, a 4k
side can transfer some pages in MD for the same transfer.
This causes a situation where MD has a different amount of
data for transfer on both sides, and the transfer is hung.
Use 4k virtual alignment for the transfer to synchronize
both sides.

WC-bug-id: LU-19180
Fixes: 262957eaef ("LU-18691 quota: quota interop check for 64k page clients")
Signed-off-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I2848b3d6fd0d1e2cf014c7130d7e75007eb63d7c
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163377
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Reviewed-by: Rajeev Mishra <rajeevm@hpe.com>