LUS-13080 lov: use-after-free of lov_stripe_md
With the conversion to kref() a LASSERT() that prevented a UAF
was removed. However some callers of lsm_addref() handled this
case gracefully so the LASSERT() was not in the correct location.

Remove the UAF race with kref_get_unless_zero() and returning NULL.
Callers which cannot gracefully handle the UAF path should LASSERT()

Fixes: c3fbc7d721 ("LU-16796 lov: Change struct lov_stripe_md to use kref")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I362c188cdd589114f21f1929f2c95d16114af811
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163459
Reviewed-by: Christopher Horn <chris.horn@hpe.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>