LUS-12962 ldiskfs: check fast symlink for ea_inode correctly
The check for a fast symlink in the presence of only an external
xattr inode is incorrect. If a fast symlink has no xattr block
(i_file_acl == 0), but has an external xattr inode that increases
inode i_blocks, then the check for a fast symlink will incorrectly
fail and __ext4_iget()->ext4_ind_check_inode() will report an error
when the inode is next loaded into memory:

    # ln -s foo /mnt/tmp/bar
    # setfattr -h -n trusted.test \
               -v "$(head -200 /etc/services)" /mnt/tmp/bar
    # umount /mnt/tmp
    # mount /mnt/tmp
    # ls -l /mnt/tmp
    ls: cannot access '/mnt/tmp/bar': Structure needs cleaning
    total 4
     ? l?????????? ? ?    ?        ?            ? bar
    # dmesg | tail -1
    EXT4-fs error (device dm-8): __ext4_iget:5098:
        inode #24578: block 7303014: comm ls: invalid block

The ext4_inode_is_fast_symlink() check should use the superblock
EA_INODE feature flag, not the inode EXT4_EA_INODE_FL flag, since
the latter is only set on the xattr inode itself, and not on the
inode that references it.

Lustre-change: https://review.whamcloud.com/59879
Lustre-commit: d0c75da783013e61c6c53bdd26337d72b3679961
Lustre-author: Andreas Dilger <adilger@whamcloud.com>

WC-bug-id: LU-19121
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ia6dc3b50bd2e7ce539ffb8a5d647d719e0fc7ea5
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163471
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@hpe.com>
Reviewed-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Tested-by: Vitaly Fertman <vitaly.fertman@hpe.com>