LUS-13084 lnet: Fix MR forwarding
The LUS-12539 change https://es-gerrit.hpc.amslabs.hpecorp.net/163191
introduced a regression in the multi-rail forwarding code. MR
forwarding allows a router to select a different target interface for
a message when the one chosen by the sender is down or unhealthy.

Prior to LUS-12539, lnet_is_peer_ni_alive() would check both the down
and unhealthy criteria, but now it only checks the down criteria. We
need to add a check for the peer NI's health value.

There was also a second bug here where we were inadvertently checking
the MR capability of the destination peer rather than the source peer.
This is also fixed.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Idaa40618d1df4c6793266c7cf9ae0d63a5a8bff7
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163521
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Reviewed-by: Alexander Boyko <alexander.boyko@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>