LUS-13096 ldlm: recovery_done is to be used after recovery
ldlm_reprocess_recovery_done() is to be used only after recovery as
it tries to reprocess all the locks in the waiting queues. In other
cases LDLM_PROCESS_RECOVERY is excessive and LDLM_PROCESS_RESCAN is
to be used instead - what is usually called on a lock cancel.

This reduces the amount of time spent on reprocessing.

Fixes: 2dc19f20ba ("LU-7372 mgs: reprocess all locks at device fini")
Fixes: 05e6ccd344 ("LU-13456 ldlm: fix reprocessing of locks with more bits")
Signed-off-by: Vitaly Fertman <vitaly.fertman@hpe.com>
Change-Id: I78bdb8a471e0662c53b2557533e7243f3793c192
Reviewed-on: https://es-gerrit.hpc.amslabs.hpecorp.net/163540
Reviewed-by: Andriy Skulysh <andriy.skulysh@hpe.com>
Reviewed-by: Alexey Lyashkov <alexey.lyashkov@hpe.com>
Tested-by: Alexander Lezhoev <alexander.lezhoev@hpe.com>